Fix child handling. Cleanup. Use QSpiObjectReference by value.
Passing around references of QSpiObjectReference is ugly.
The class could be implemented as shared data class if we want to optimize later.
Give it a default ctor that creates a null reference.

Cleanup some whitespaces.

Fix several instances where we did not refer to the child paremeter.
Fix child count and implement a getChild to properly handle finding of child objects.

Add some more widgets to the simple sample.