Fix VisibleDataChanged event.
atspi returns "VisibledataChanged" rather then "VisibleDataChanged" (lower-case
d) and since we do a case-sensitive check for the name we failed to detect
the proper VisibleDataChanged.

This is fixed now by just checking for both. I chosed to support both case-styles
cause imho its a typo in atspi that may be fixed some day and if that happens
we already support the proper/new name plus the current/old one.